Intital commit for GET/POST
This is the first commit where I had been working on getting a response from the sandbox API for the Coinbase pro page. We have GET functions working to get account info if we wanted and then we also have the POST orders function that actually creates buy orders on the site.