Buy/Sell Via Order
We can now buy and sell any given currency, it is also improved from the old price method, we now use the keyword funds in the json object which matches the amount of fiat you want to buy or sell for the given crypto. This has a higher takers fee so in future iterations it might be good to use the limit and a websocket feed to get current price manually.